+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Support data attributes for the back-link component (PR #773)
 * Create shared helper for components (PR #759)
 * Use delegated event handlers for checkbox events (PR #770)
+* Add margin option to details component (PR #768)
 
 ## 16.1.0